refactor: extracted functionality into functions for better readibility
